finish all the requirments with small 🐞 in ShoeListFragment.kt and gravity of logout_menu.xml

+++ app/src/main/java/com/aziza/shoestore/model/Shoe.kt
@@ -6,7 +6,7 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 data class Shoe(
     var name: String,
-    var size: Double,
+    var size: Int,
     var company: String,
     var description: String,
     val images: List<String> = mutableListOf()
